DB-5662: make sure BackupEndpointObserver is up to date for all platforms
