+++ hbase_sql/hbase1.0.0.x/src/main/java/com/splicemachine/hbase/BackupEndpointObserver.java
@@ -20,12 +20,8 @@ import com.splicemachine.access.HConfiguration;
 import com.splicemachine.backup.BackupRestoreConstants;
 import com.splicemachine.backup.BackupUtils;
 import com.splicemachine.coprocessor.SpliceMessage;
-import com.splicemachine.db.iapi.error.StandardException;
-import com.splicemachine.pipeline.Exceptions;
-import com.splicemachine.si.constants.SIConstants;
 import com.splicemachine.utils.SpliceLogUtils;
 import org.apache.hadoop.conf.Configuration;
-import org.apache.hadoop.fs.FSDataOutputStream;
 import org.apache.hadoop.fs.FileSystem;
 import org.apache.hadoop.fs.Path;
 import org.apache.hadoop.hbase.Coprocessor;
@@ -34,12 +30,9 @@ import org.apache.hadoop.hbase.coprocessor.CoprocessorService;
 import org.apache.hadoop.hbase.coprocessor.ObserverContext;
 import org.apache.hadoop.hbase.coprocessor.RegionCoprocessorEnvironment;
 import org.apache.hadoop.hbase.regionserver.*;
-import org.apache.hadoop.hbase.util.Bytes;
 import org.apache.hadoop.hbase.util.FSUtils;
-import org.apache.hadoop.hbase.zookeeper.RecoverableZooKeeper;
 import org.apache.log4j.Logger;
 import org.apache.zookeeper.CreateMode;
-import org.apache.zookeeper.KeeperException;
 import org.apache.zookeeper.ZooDefs;
 
 import java.io.IOException;
@@ -108,8 +101,8 @@ public class BackupEndpointObserver extends BackupBaseRegionObserver implements 
                 responseBuilder.setReadyForBackup(false);
             } else {
                 // A region might have been in backup
-                if (!regionIsBeingBackup()) {
-                   canceled = backupCanceled();
+                if (!BackupUtils.regionIsBeingBackup(tableName, regionName, path)) {
+                   canceled = BackupUtils.backupCanceled();
                     if (!canceled) {
                         HBasePlatformUtils.flush(region);
                         // Create a ZNode to indicate that the region is being copied
@@ -152,7 +145,7 @@ public class BackupEndpointObserver extends BackupBaseRegionObserver implements 
         if (LOG.isDebugEnabled())
             SpliceLogUtils.debug(LOG, "BackupEndpointObserver.preSplit()");
 
-        waitForBackupToComplete();
+        BackupUtils.waitForBackupToComplete(tableName, regionName, path);
         isSplitting = true;
         super.preSplit(e);
     }
@@ -169,7 +162,7 @@ public class BackupEndpointObserver extends BackupBaseRegionObserver implements 
     public void preCompactSelection(ObserverContext<RegionCoprocessorEnvironment> c, Store store, List<StoreFile> candidates) throws IOException {
         if (LOG.isDebugEnabled())
             SpliceLogUtils.debug(LOG, "BackupEndpointObserver.preCompactSelection()");
-        waitForBackupToComplete();
+        BackupUtils.waitForBackupToComplete(tableName, regionName, path);
         super.preCompactSelection(c, store, candidates);
     }
 
@@ -178,7 +171,7 @@ public class BackupEndpointObserver extends BackupBaseRegionObserver implements 
         if (LOG.isDebugEnabled())
             SpliceLogUtils.debug(LOG, "BackupEndpointObserver.preCompact()");
 
-        waitForBackupToComplete();
+        BackupUtils.waitForBackupToComplete(tableName, regionName, path);
         return super.preCompact(e, store, scanner, scanType);
     }
 
@@ -186,7 +179,7 @@ public class BackupEndpointObserver extends BackupBaseRegionObserver implements 
     public void preFlush(ObserverContext<RegionCoprocessorEnvironment> e) throws IOException {
         if (LOG.isDebugEnabled())
             SpliceLogUtils.debug(LOG, "BackupEndpointObserver.preFlush()");
-        waitForBackupToComplete();
+        BackupUtils.waitForBackupToComplete(tableName, regionName, path);
         super.preFlush(e);
     }
 
@@ -197,135 +190,10 @@ public class BackupEndpointObserver extends BackupBaseRegionObserver implements 
         try {
             if (namespace.compareTo("splice") != 0)
                 return;
-            captureIncrementalChanges(resultFile.getPath().getName());
-        }
-        catch (Exception ex) {
+            BackupUtils.captureIncrementalChanges(conf, region, path, fs, rootDir, backupDir,
+                    tableName, resultFile.getPath().getName(), preparing);
+        } catch (Exception ex) {
             throw new IOException(ex);
-        }
-    }
-
-    private void waitForBackupToComplete() throws IOException{
-        int i = 0;
-        while (regionIsBeingBackup()) {
-            try {
-                if (LOG.isDebugEnabled())
-                    SpliceLogUtils.debug(LOG, "wait for backup to complete");
-                Thread.sleep(100 * (long) Math.pow(2, i));
-                i++;
-            }
-            catch (InterruptedException e) {
-                throw new IOException(e);
-            }
-        }
-    }
-
-    private boolean regionIsBeingBackup() {
-        boolean isBackup = false;
-        try {
-            RecoverableZooKeeper zooKeeper = ZkUtils.getRecoverableZooKeeper();
-            if (zooKeeper.exists(path, false) == null) {
-                if (LOG.isDebugEnabled())
-                    SpliceLogUtils.debug(LOG, "Table %s region %s is not in backup", tableName, regionName);
-                isBackup = false;
-            }
-            else {
-                byte[] status = ZkUtils.getData(path);
-                if (Bytes.compareTo(status, HConfiguration.BACKUP_IN_PROGRESS) == 0) {
-                    if (LOG.isDebugEnabled())
-                        SpliceLogUtils.debug(LOG, "Table %s region %s is in backup", tableName, regionName);
-                    isBackup = true;
-                }
-                else if (Bytes.compareTo(status, HConfiguration.BACKUP_DONE) == 0) {
-                    if (LOG.isDebugEnabled())
-                        SpliceLogUtils.debug(LOG, "Table %s region %s is done with backup", tableName, regionName);
-                    isBackup = false;
-                }
-                else {
-                    throw new RuntimeException("Unexpected data in node:" + path);
-                }
-            }
-        } catch (Exception e) {
-            throw new RuntimeException(e);
-        }
-
-        return isBackup;
-    }
-
-    private boolean backupCanceled() throws KeeperException, InterruptedException {
-        RecoverableZooKeeper zooKeeper = ZkUtils.getRecoverableZooKeeper();
-        String path = HConfiguration.getConfiguration().getBackupPath();
-        return zooKeeper.exists(path, false) == null;
-    }
-
-    /**
-     * An HFile is eligible for incremental backup if
-     * 1) There is an ongoing full backup, flush is not triggered by preparing and backup for this region is done.
-     * 2) There is no ongoing backup, AND there is a previous full/incremental backup
-     * 3) There is an ongoing incremental backup
-     * @param fileName
-     * @throws StandardException
-     */
-    private void captureIncrementalChanges(String fileName) throws StandardException {
-        boolean shouldRegister = false;
-        try {
-            RecoverableZooKeeper zooKeeper = ZkUtils.getRecoverableZooKeeper();
-            String spliceBackupPath = HConfiguration.getConfiguration().getBackupPath();
-            if (zooKeeper.exists(spliceBackupPath, false) != null) {
-                byte[] backupType = ZkUtils.getData(spliceBackupPath);
-                if (Bytes.compareTo(backupType, BackupRestoreConstants.BACKUP_TYPE_FULL_BYTES) == 0) {
-                    if (!preparing && zooKeeper.exists(path, false) != null) {
-                        byte[] status = ZkUtils.getData(path);
-                        if (Bytes.compareTo(status, HConfiguration.BACKUP_DONE) == 0) {
-                            if (LOG.isDebugEnabled()) {
-                                SpliceLogUtils.debug(LOG, "Table %s is being backup", tableName);
-                            }
-                            shouldRegister = true;
-                        }
-                    }
-                }
-                else {
-                    shouldRegister = true;
-                }
-
-            }
-            else if (BackupUtils.existsDatabaseBackup(fs, rootDir)) {
-                shouldRegister = true;
-            }
-            if (shouldRegister) {
-                registerHFile(fileName);
-            }
-        }
-        catch (Exception e) {
-            e.printStackTrace();
-            throw Exceptions.parseException(e);
-        }
-    }
-
-    /**
-     * Register this HFile for incremental backup by creating an empty file
-     * backup/data/splice/tableName/regionName/V/fileName
-     * @param fileName
-     * @throws StandardException
-     */
-    private void registerHFile(String fileName) throws StandardException {
-
-        FSDataOutputStream out = null;
-        try {
-            if (!fs.exists(new Path(backupDir, BackupRestoreConstants.REGION_FILE_NAME))) {
-                HRegionFileSystem.createRegionOnFileSystem(conf, fs, backupDir.getParent(), region.getRegionInfo());
-            }
-            out = fs.create(new Path(backupDir.toString() + "/" + SIConstants.DEFAULT_FAMILY_NAME + "/" + fileName));
-        }
-        catch (Exception e) {
-            throw Exceptions.parseException(e);
-        }
-        finally {
-            try {
-                out.close();
-            }
-            catch (Exception e) {
-                throw Exceptions.parseException(e);
-            }
         }
     }
 }
